♻️ Extract methods (#1140)
